requirements.txt
google-api-python-client==2.134.0
google-auth-oauthlib==1.2.0
google-cloud-pubsub==2.20.0
slack-sdk==3.35.0
transformers==4.40.0
torch==2.3.0
python-dotenv==1.0.1
pytest==7.4.3
pytest-cov==4.1.0

# Optional: For a web server if handling Pub/Sub push via HTTP endpoint
# flask==3.0.3
# fastapi==0.111.0
# uvicorn[standard]==0.29.0 # For running FastAPI 